ref(metrics): Refactor tests into their own modules
